Move run_editor to utils
